pyproject.toml
[tool.poetry]
name = "FYP"
version = "0.1.0"
description = "Final Year Project"
authors = ["Michael Strauss <michael.strauss@live.com.au>"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
tqdm = "^4.48.2"
requests = "^2.24.0"
loguru = "^0.5.2"
torch = "^1.6.0"
torchvision = "^0.7.0"
pytorch-lightning = "^0.9.0"
matplotlib = "^3.3.1"
opencv-python = "^4.4.0"
scipy = "^1.5.2"
comet-ml = "^3.2.0"
python-dotenv = "^0.14.0"
pytorch-nlp = "^0.5.0"
lru-dict = "^1.1.6"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^20.8b1"
pylint = "^2.6.0"
jupyter = "^1.0.0"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
